pyproject.toml
[tool.poetry]
name = "Base FastApi project"
version = "0.1.0"
description = ""
authors = ["RamilAbzalov <ramil.abzalov@scalewill.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.0"
python-multipart = "^0.0.5"
email-validator = "^1.1.3"
requests = "^2.27.1"
passlib = "^1.7.4"
SQLAlchemy = "^1.4.32"
alembic = "^1.7.7"
Jinja2 = "^3.0.3"
boto3 = "^1.21.22"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^0.20.0"
strawberry-graphql = "^0.104.0"
sqlmodel = "^0.0.6"
bcrypt = "^3.2.0"
PyJWT = "^2.3.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
inflection = "^0.5.1"
gunicorn = "^20.1.0"
asyncpg = "^0.25.0"
psycopg2-binary = "^2.9.3"
greenlet = "^1.1.2"
databases = {extras = ["postgresql"], version = "^0.5.5"}

[tool.poetry.dev-dependencies]
isort = "^4.3.21"
sqlalchemy-stubs = "^0.3"
pytest-cov = "^2.8.1"
mypy = "^0.942"
flake8 = "^4.0.1"
autopep8 = "^1.6.0"
wemake-python-styleguide = "^0.16.1"
pytest = "^7.1.1"
ipython = "^8.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
